correct constants and cli options

+++ src/constants.jl
@@ -1,2 +1,2 @@
-const max_excel_columns = 200
-const max_excel_lines = 1000
+const max_excel_columns::Int = 200
+const max_excel_lines::Int = 1000
